Fix code scanning alert

+++ src/c/profile/transport/stream_framing/stream_framing_protocol.c
@@ -169,7 +169,7 @@ size_t uxr_write_framed_msg(
     uint16_t written_len = 0;
     uint16_t crc = 0;
     bool cond = true;
-    while (written_len < len && cond)
+    while ((written_len < (uint16_t) len) && cond)
     {
         octet = *(buf + written_len);
         if (uxr_add_next_octet(framing_io, octet))
